% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.design}
\alias{autoplot.design}
\title{Plot the layout of an experimental design}
\usage{
\method{autoplot}{design}(
  object,
  rotation = 0,
  size = 4,
  margin = FALSE,
  palette = "default",
  row = NULL,
  column = NULL,
  block = NULL,
  treatments = NULL,
  legend = TRUE,
  ...
)
}
\arguments{
\item{object}{A \code{design} object, as produced by \code{\link[=design]{design()}}.}

\item{rotation}{Rotate the treatment labels within the plot. Allows for easier reading of long treatment labels. Number between 0 and 360 (inclusive) - default 0}

\item{size}{Increase or decrease the text size within the plot for treatment labels. Numeric with default value of 4.}

\item{margin}{Logical (default \code{FALSE}). A value of \code{FALSE} will expand the plot to the edges of the plotting area i.e. remove white space between plot and axes.}

\item{palette}{A string specifying the colour scheme to use for plotting or a vector of custom colours to use as the palette. Default is equivalent to "Spectral". Colour blind friendly palettes can also be provided via options \code{"colour blind"} (or \code{"colour blind"}, both equivalent to \code{"viridis"}), \code{"magma"}, \code{"inferno"}, \code{"plasma"}, \code{"cividis"}, \code{"rocket"}, \code{"mako"} or \code{"turbo"}. Other palettes from \code{\link[scales:brewer_pal]{scales::brewer_pal()}} are also possible.}

\item{row}{A variable to plot a column from \code{object} as rows.}

\item{column}{A variable to plot a column from \code{object} as columns.}

\item{block}{A variable to plot a column from \code{object} as blocks.}

\item{treatments}{A variable to plot a column from \code{object} as treatments.}

\item{legend}{Logical (default \code{TRUE}). If \code{TRUE}, displays the legend for treatment colours.}

\item{...}{Arguments passed to \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}} for the plot labels.}
}
\value{
A \code{ggplot2} object.
}
\description{
Produces a plot of the plot/field layout for a \code{\link[=design]{design()}} result, with plots
coloured by treatment and block boundaries drawn for blocked designs.
}
\examples{
des.out <- design(type = "crd", treatments = c(1, 5, 10, 20),
                  reps = 5, nrows = 4, ncols = 5, seed = 42, plot = FALSE)
autoplot(des.out)

# Colour blind friendly colours
autoplot(des.out, palette = "colour-blind")

# Alternative colour scheme
autoplot(des.out, palette = "plasma")

# Custom colour palette
autoplot(des.out, palette = c("#ef746a", "#3fbfc5", "#81ae00", "#c37cff"))

# Visualise different components of a split plot design
des.out <- design(type = "split", treatments = c("A", "B"), sub_treatments = 1:4,
reps = 4, nrows = 8, ncols = 4, brows = 4, bcols = 2, seed = 42)

# Show the wholeplot components
autoplot(des.out, treatments = wholeplots)

# Display block level
autoplot(des.out, treatments = block)
}
\seealso{
\code{\link[=design]{design()}}
}
