% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resplot.R
\name{combine_plots}
\alias{combine_plots}
\title{Combine plots into final output}
\usage{
combine_plots(plots, shapiro_result, model_call, call, call.size, label.size)
}
\arguments{
\item{plots}{List of ggplot objects}

\item{shapiro_result}{Result from Shapiro test or NULL}

\item{model_call}{Model call string or NULL}

\item{call}{Logical, whether to include call}

\item{call.size}{Size of call text}

\item{label.size}{Size of labels}
}
\description{
Combine plots into final output
}
\keyword{internal}
