% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare_version}
\alias{compare_version}
\title{Function to compare package version for mocking}
\usage{
compare_version(a, b)
}
\arguments{
\item{a, b}{Character strings representing package version numbers.}
}
\value{
Numeric. \code{0} if the numbers are equal, \code{-1} if \code{b} is later and \code{1} if \code{a} is later
}
\description{
Function to compare package version for mocking
}
\keyword{internal}
