% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_asreml.R
\name{compare_versions}
\alias{compare_versions}
\title{Compare installed version of ASReml-R with available versions}
\usage{
compare_versions()
}
\value{
TRUE if a newer version is available online, FALSE otherwise
}
\description{
Compare installed version of ASReml-R with available versions
}
\keyword{internal}
