% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resplot.R
\name{create_diagnostic_plots}
\alias{create_diagnostic_plots}
\title{Create diagnostic plots}
\usage{
create_diagnostic_plots(group_residuals, res.type, axes.size, label.size)
}
\arguments{
\item{group_residuals}{Data frame with residuals and fitted values}

\item{axes.size}{Size of axes labels}

\item{label.size}{Size of plot labels}
}
\description{
Create diagnostic plots
}
\keyword{internal}
