% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_excel.R
\name{.export_to_excel}
\alias{.export_to_excel}
\title{Export design and layout data to a colour-formatted Excel file}
\usage{
.export_to_excel(design_df, layout_df, filename, value_column, palette)
}
\arguments{
\item{design_df}{A data frame containing the experimental design (raw data).}

\item{layout_df}{A data frame representing the layout matrix of treatments.}

\item{filename}{The name of the Excel file to be created.}

\item{value_column}{The name of the column in \code{design_df} indicating treatment.}

\item{palette}{A palette name or vector of colours used for treatments.}
}
\description{
Internal function to export experimental layout and raw data to an Excel file
with treatment-based colour formatting.
}
\keyword{internal}
