% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_asreml.R
\name{download_asreml_package}
\alias{download_asreml_package}
\title{Download asreml package file}
\usage{
download_asreml_package(url, verbose = FALSE)
}
\arguments{
\item{verbose}{Logical for verbose output}
}
\description{
Download asreml package file
}
\keyword{internal}
