% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldlayout.R
\name{field_layout}
\alias{field_layout}
\title{Field layout}
\usage{
field_layout(data, row_var, col_var, other_var)
}
\arguments{
\item{data}{A data frame with the field values.}

\item{row_var}{The row (y axis) variable.}

\item{col_var}{The column (x axis) variable.}

\item{other_var}{Another variable to use for breaking up the plotting area such as Blocks or Wholeplots}
}
\value{
A ggplot2 plot
}
\description{
Easily create plots of field layouts.
}
