% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resplot.R
\name{format_output_resplot}
\alias{format_output_resplot}
\title{Format final output based on facet structure}
\usage{
format_output_resplot(output, facet, facet_name, onepage, onepage_cols)
}
\arguments{
\item{output}{List of plot objects}

\item{facet}{Number of facets}

\item{facet_name}{Names of facets}

\item{onepage}{Logical, combine plots on one page}

\item{onepage_cols}{Number of columns for onepage layout}
}
\description{
Format final output based on facet structure
}
\keyword{internal}
