% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_methods.R
\name{get_predictions.lm}
\alias{get_predictions.lm}
\title{Get Predictions for Linear Models}
\usage{
\method{get_predictions}{lm}(model.obj, classify, ...)
}
\arguments{
\item{model.obj}{A linear model object.}

\item{classify}{Name of predictor variable as a string.}

\item{...}{Additional arguments passed to \code{emmeans::emmeans}.}
}
\value{
A list containing predictions, standard errors, degrees of freedom,
response variable label, and aliased names.
}
\description{
An S3 method for getting predictions from linear models.
}
