% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_asreml.R
\name{get_r_os}
\alias{get_r_os}
\title{Detect operating system and construct OS/version key}
\usage{
get_r_os()
}
\value{
A list with os_ver, os, ver, and arm
}
\description{
Determines OS, OS version (if applicable), architecture,
and combines this with the R version for downstream
download logic.
}
\keyword{internal}
