% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_asreml.R
\name{get_r_version_compact}
\alias{get_r_version_compact}
\title{Get R major-minor version without dot}
\usage{
get_r_version_compact()
}
\value{
Character scalar R version
}
\description{
Converts R version to a compact form (e.g. 4.4.x -> "44").
}
\keyword{internal}
