% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_asreml.R
\name{install_asreml_package}
\alias{install_asreml_package}
\title{Install the ASReml package}
\usage{
install_asreml_package(save_file, library, quiet, os, verbose = FALSE)
}
\arguments{
\item{save_file}{Path to package file}

\item{library}{Library path}

\item{quiet}{Whether to suppress messages}

\item{os}{Operating system}

\item{verbose}{Logical for verbose output}
}
\value{
TRUE if successful, FALSE otherwise
}
\description{
Install the ASReml package
}
\keyword{internal}
