% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_excel.R
\name{int2col}
\alias{int2col}
\title{Convert column number to Excel column letter}
\usage{
int2col(num)
}
\arguments{
\item{num}{Integer column number}
}
\value{
Character Excel column letter(s)
}
\description{
Convert column number to Excel column letter
}
\keyword{internal}
