% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_asreml.R
\name{is_arm}
\alias{is_arm}
\title{Detect whether the system is ARM-based}
\usage{
is_arm()
}
\value{
Logical indicating ARM architecture
}
\description{
x86_64 is treated as the implicit default.
}
\keyword{internal}
