% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_template.R
\name{list_templates}
\alias{list_templates}
\title{List available biometryassist templates}
\usage{
list_templates()
}
\value{
\code{character()} Vector of available template file names.
}
\description{
\code{list_templates()} returns a character vector of available analysis templates
included with the biometryassist package.
}
\examples{
# See what templates are available
list_templates()

}
\seealso{
\code{\link[=use_template]{use_template()}} to copy and use a template
}
