% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_asreml.R
\name{manage_file}
\alias{manage_file}
\title{Manage the downloaded file}
\usage{
manage_file(save_file, keep_file, filename, verbose = FALSE)
}
\arguments{
\item{save_file}{Path to the downloaded file}

\item{keep_file}{Whether/where to keep the file}

\item{filename}{Original filename}

\item{verbose}{Logical for verbose output}
}
\value{
logical; TRUE if file successfully handled, FALSE otherwise
}
\description{
Manage the downloaded file
}
\keyword{internal}
