% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_asreml.R
\name{map_linux_target}
\alias{map_linux_target}
\title{Map Linux distro to supported download target}
\usage{
map_linux_target(info)
}
\description{
Map Linux distro to supported download target
}
\keyword{internal}
