% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_asreml.R
\name{newer_version}
\alias{newer_version}
\title{Compare installed version of ASReml-R with available versions}
\usage{
newer_version()
}
\value{
TRUE if a newer version is available online, FALSE otherwise
}
\description{
Compare installed version of ASReml-R with available versions
}
\keyword{internal}
