% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_asreml.R
\name{parse_version_table}
\alias{parse_version_table}
\title{Parse version table from web scraping}
\usage{
parse_version_table(tables, headers)
}
\arguments{
\item{tables}{List of table data}

\item{headers}{Header information}
}
\value{
Combined data frame of version information
}
\description{
Parse version table from web scraping
}
\keyword{internal}
