% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mct.R
\name{print.mct}
\alias{print.mct}
\title{Print output of multiple_comparisons}
\usage{
\method{print}{mct}(x, decimals = 2, ...)
}
\arguments{
\item{x}{An mct object to print to the console.}

\item{decimals}{Number of decimal places to display. Default is 2.}

\item{...}{Other arguments passed to print.data.frame}
}
\value{
The original object invisibly.
}
\description{
Print output of multiple_comparisons
}
\examples{
dat.aov <- aov(Petal.Width ~ Species, data = iris)
output <- multiple_comparisons(dat.aov, classify = "Species")
print(output)
print(output, decimals = 4)
}
\seealso{
\code{\link[=multiple_comparisons]{multiple_comparisons()}}
}
