% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{quiet}
\alias{quiet}
\title{Function to suppress output if desired, especially useful for ASReml output}
\usage{
quiet(x)
}
\arguments{
\item{x}{A function call with output to be suppressed.}
}
\value{
The invisible output of the function called.
}
\description{
Function to suppress output if desired, especially useful for ASReml output
}
\keyword{internal}
