% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_asreml.R
\name{remove_existing_asreml}
\alias{remove_existing_asreml}
\title{Remove existing ASReml installation}
\usage{
remove_existing_asreml()
}
\arguments{
\item{os}{Operating system}
}
\description{
Remove existing ASReml installation
}
\keyword{internal}
