% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{residuals.aovlist}
\alias{residuals.aovlist}
\title{Output residuals from a multi-stratum ANOVA}
\usage{
\method{residuals}{aovlist}(model.obj)
}
\arguments{
\item{model.obj}{An \code{aovlist} model object.}
}
\description{
Internal helper function calculate the residuals from a multi-stratum ANOVA implemented using aov.
}
\keyword{internal}
