% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resplot.R
\name{resplot}
\alias{resplot}
\title{Produce residual plots of linear models}
\usage{
resplot(
  model.obj,
  res.type = "standardised",
  shapiro = TRUE,
  call = FALSE,
  label.size = 10,
  axes.size = 10,
  call.size = 9,
  onepage = FALSE,
  onepage_cols = 3,
  mod.obj
)
}
\arguments{
\item{model.obj}{An \code{aov}, \code{aovlist}, \code{lm}, \code{lme} (\code{\link[nlme:lme]{nlme::lme()}}), \code{lmerMod} (\code{\link[lme4:lmer]{lme4::lmer()}}), \code{asreml} or \code{mmer} (sommer) model object.}

\item{res.type}{A character value for whether the residuals displayed are \code{simple} or \code{standardised}. Default is \code{standardised}.}

\item{shapiro}{(Logical) Display the Shapiro-Wilk test of normality on the plot? This test is unreliable for larger numbers of observations and will not work with n >= 5000 so will be omitted from any plots.}

\item{call}{(Logical) Display the model call on the plot?}

\item{label.size}{A numeric value for the size of the label (A,B,C) font point size.}

\item{axes.size}{A numeric value for the size of the axes label font size in points.}

\item{call.size}{A numeric value for the size of the model displayed on the plot.}

\item{onepage}{(Logical) If TRUE and there are multiple plots, combines up to 6 plots per page.}

\item{onepage_cols}{Integer. Number of columns to use in grid layout when onepage=TRUE. Default is 3.}

\item{mod.obj}{Deprecated to be consistent with other functions. Please use \code{model.obj} instead.}
}
\value{
A ggplot2 object containing the diagnostic plots.
}
\description{
Produces plots of residuals for assumption checking of linear (mixed) models.
}
\examples{
dat.aov <- aov(Petal.Length ~ Petal.Width, data = iris)
resplot(dat.aov)
resplot(dat.aov, call = TRUE)
}
