% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satab.R
\name{satab}
\alias{satab}
\title{Produces a skeletal ANOVA table}
\usage{
satab(design.obj)
}
\arguments{
\item{design.obj}{An \code{agricolae} design object.}
}
\value{
Prints skeletal ANOVA table to console output.
}
\description{
Produces a skeletal ANOVA table
}
\keyword{internal}
