% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resplot.R
\name{shapiro_test}
\alias{shapiro_test}
\title{Shapiro-Wilk test}
\usage{
shapiro_test(group_residuals)
}
\arguments{
\item{group_residuals}{Data frame with residuals}
}
\description{
Shapiro-Wilk test
}
\keyword{internal}
