% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_graph.R
\name{summary_graph}
\alias{summary_graph}
\title{Summary Graph}
\usage{
summary_graph(data, response, exp_var, resp_units, ...)
}
\arguments{
\item{data}{A data frame containing the variables to be plotted.}

\item{response}{The response variable to plot.}

\item{exp_var}{The explanatory (or grouping) variable to plot.}

\item{resp_units}{A string providing units to display on the response variable axis.}
}
\value{
A ggplot2 plot object
}
\description{
Produce a graphical summary of variables from a data frame. Variables are plotted
as box plots, grouped and, if enough variables are provided, coloured and facetted.
}
\examples{

summary_graph(iris, "Petal.Length", "Species", "mm")
summary_graph(iris, Petal.Length, "Species", "mm")
summary_graph(npk, "yield", c("N", "P", "K"), "lb/plot")

}
