% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_buffers.R
\name{add_buffers}
\alias{add_buffers}
\title{Add buffers to an existing design}
\usage{
add_buffers(design_obj, type)
}
\arguments{
\item{design_obj}{A design object (with class "design") from the design() function}

\item{type}{The type of buffer to add}
}
\value{
The modified design object with buffers added
}
\description{
Add buffers to an existing design
}
