% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot}
\alias{autoplot}
\title{Generate automatic plots for objects generated in biometryassist}
\usage{
autoplot(object, ...)
}
\arguments{
\item{object}{An object created by \code{biometryassist}. Methods are provided for
the \code{mct}, \code{design}, \code{pairwise_comparisons} and \code{reference_comparisons}
classes.}

\item{...}{Arguments passed to the individual \code{autoplot} methods.}
}
\value{
A \code{ggplot2} object.
}
\description{
\code{\link[ggplot2:autoplot]{ggplot2::autoplot()}} methods are provided for the objects created by
\code{biometryassist}. See the per-class methods for the available options:
\code{\link[=autoplot.mct]{autoplot.mct()}} for \code{\link[=multiple_comparisons]{multiple_comparisons()}} output, \code{\link[=autoplot.design]{autoplot.design()}} for
\code{\link[=design]{design()}} output, and \code{\link[=autoplot.pairwise_comparisons]{autoplot.pairwise_comparisons()}} /
\code{\link[=autoplot.reference_comparisons]{autoplot.reference_comparisons()}}, which are documented alongside their
respective functions.
}
\seealso{
\code{\link[=autoplot.mct]{autoplot.mct()}}, \code{\link[=autoplot.design]{autoplot.design()}}, \code{\link[=multiple_comparisons]{multiple_comparisons()}} and \code{\link[=design]{design()}}
}
