% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biometryassist-deprecated.R, R/resplot.R
\name{biometryassist-deprecated}
\alias{biometryassist-deprecated}
\alias{resplt}
\title{Deprecated functions in package \pkg{biometryassist}.}
\usage{
resplt(
  model.obj,
  shapiro = TRUE,
  call = FALSE,
  label.size = 10,
  axes.size = 10,
  call.size = 9,
  mod.obj
)
}
\value{
No return value.

A list containing ggplot2 objects which are diagnostic plots.
For \code{resplt}, use \code{\link[=resplot]{resplot()}}.
}
\description{
The functions listed below are deprecated and will be removed in
the future. When possible, alternative functions with similar
functionality are also mentioned. Help pages for deprecated functions are
available at \code{help("<function>-deprecated")}.
}
\section{resplt}{

Residual plots of linear models.
}

\keyword{internal}
