% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biometryassist-package.R
\docType{package}
\name{biometryassist}
\alias{biometryassist}
\alias{_PACKAGE}
\alias{biometryassist-package}
\title{biometryassist: Functions to Assist Design and Analysis of Agronomic Experiments}
\description{
Provides functions to aid in the design and analysis of agronomic and agricultural experiments through easy access to documentation and helper functions, especially for users who are learning these concepts. While not required for most functionality, this package enhances the `asreml` package which provides a computationally efficient algorithm for fitting mixed models using Residual Maximum Likelihood. It is a commercial package that can be purchased as 'asreml-R' from 'VSNi' \url{https://vsni.co.uk/}, who will supply a zip file for local installation/updating (see \url{https://asreml.kb.vsni.co.uk/}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://biometryhub.github.io/biometryassist/}
  \item Report bugs at \url{https://github.com/biometryhub/biometryassist/issues}
}

}
\author{
\strong{Maintainer}: Sam Rogers \email{biometrytraining@adelaide.edu.au}

Authors:
\itemize{
  \item Sharon Nielsen \email{biometrytraining@adelaide.edu.au}
  \item Annie Conway \email{biometrytraining@adelaide.edu.au}
}

Other contributors:
\itemize{
  \item University of Adelaide (https://adelaide.edu.au/) [copyright holder, funder]
  \item Grains Research and Development Corporation (https://grdc.com.au/) [copyright holder, funder]
}

}
\keyword{internal}
