% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biometryassist-package.R
\docType{package}
\name{biometryassist}
\alias{biometryassist}
\alias{_PACKAGE}
\alias{biometryassist-package}
\title{biometryassist: Functions to Assist Design and Analysis of Agronomic Experiments}
\description{
Provides functions to aid in the design and analysis of agronomic and agricultural experiments through easy access to documentation and helper functions, especially for users who are learning these concepts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://biometryhub.github.io/biometryassist/}
  \item Report bugs at \url{https://github.com/biometryhub/biometryassist/issues}
}

}
\author{
\strong{Maintainer}: Sam Rogers \email{biometrytraining@adelaide.edu.au}

Authors:
\itemize{
  \item Sharon Nielsen \email{biometrytraining@adelaide.edu.au}
  \item Annie Conway \email{biometrytraining@adelaide.edu.au}
}

Other contributors:
\itemize{
  \item University of Adelaide (https://adelaide.edu.au/) [copyright holder, funder]
  \item Grains Research and Development Corporation (https://grdc.com.au/) [copyright holder, funder]
}

}
\keyword{internal}
