% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_buffers.R
\name{create_buffers}
\alias{create_buffers}
\title{Create buffers for design plots}
\usage{
create_buffers(design, type, blocks = FALSE)
}
\arguments{
\item{design}{The data frame of the design.}

\item{type}{The type of buffer. One of edge, row, column, double row, double column, or block (coming soon).}

\item{blocks}{Does the design data frame contain blocks?}
}
\value{
The original data frame, updated to include buffers
}
\description{
Create buffers for design plots
}
\keyword{internal}
