% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_asreml.R
\name{create_mac_folder}
\alias{create_mac_folder}
\title{Create the folder MacOS needs for licensing}
\usage{
create_mac_folder()
}
\value{
logical; TRUE if folder successfully created, otherwise it will error
}
\description{
Create the folder MacOS needs for licensing
}
\keyword{internal}
