% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resplot_methods.R
\name{extract_model_info}
\alias{extract_model_info}
\title{Extract model information using S3 dispatch}
\usage{
extract_model_info(model.obj, call = FALSE)
}
\arguments{
\item{model.obj}{Model object}

\item{call}{Logical, whether to extract model call}
}
\description{
Extract model information using S3 dispatch
}
\keyword{internal}
