% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fitted.aovlist}
\alias{fitted.aovlist}
\title{Output the fitted values from a multi-stratum ANOVA}
\usage{
\method{fitted}{aovlist}(model.obj)
}
\arguments{
\item{model.obj}{An \code{aovlist} model object.}
}
\description{
Internal helper function calculate the fitted values from a multi-stratum ANOVA implemented using aov.
}
\keyword{internal}
