% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_methods.R
\name{get_predictions}
\alias{get_predictions}
\title{Get Predictions for Statistical Models}
\usage{
get_predictions(model.obj, classify, pred.obj = NULL, vars = NULL, ...)
}
\arguments{
\item{model.obj}{A model object.}

\item{classify}{Name of predictor variable as a string.}

\item{pred.obj}{Optional precomputed prediction object.}

\item{vars}{Variables used in the model.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
A list containing predictions, standard errors, degrees of freedom,
response variable label, and aliased names.
}
\description{
A generic function to get predictions for statistical models.
}
