% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_asreml.R
\name{get_r_os}
\alias{get_r_os}
\title{Get the version of R and OS}
\usage{
get_r_os()
}
\value{
A list with the version of R and the OS in a standard format
}
\description{
Get the version of R and OS
}
\keyword{internal}
