% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_asreml.R
\name{get_version_table}
\alias{get_version_table}
\title{Get released versions of ASReml-R in lookup table}
\usage{
get_version_table()
}
\value{
A list of data frames containing the version number and release date of released ASReml-R versions for comparison
}
\description{
Get released versions of ASReml-R in lookup table
}
\keyword{internal}
