% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{handle_deprecated_param}
\alias{handle_deprecated_param}
\title{Handle deprecated parameters}
\usage{
handle_deprecated_param(
  old_param,
  new_param = NULL,
  custom_msg = NULL,
  call_env = parent.frame()
)
}
\arguments{
\item{old_param}{Name of the deprecated parameter}

\item{new_param}{Name of the replacement parameter or NULL if parameter is being removed}

\item{custom_msg}{Optional custom message to append to the warning}

\item{call_env}{Environment where to check for the deprecated parameter}
}
\value{
Nothing, called for side effects (warnings)
}
\description{
Simple internal function to warn about deprecated parameters
}
\keyword{internal}
