% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_asreml.R
\name{manage_file}
\alias{manage_file}
\title{Manage the downloaded file}
\usage{
manage_file(save_file, keep_file, filename)
}
\value{
logical; TRUE if file successfully moved, FALSE otherwise
}
\description{
Manage the downloaded file
}
\keyword{internal}
