% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mct.R
\name{print.mct}
\alias{print.mct}
\title{Print method for multiple_comparisons}
\usage{
\method{print}{mct}(x, ...)
}
\arguments{
\item{x}{An mct object to print to the console.}

\item{...}{Other arguments to be passed through.}
}
\value{
The original object invisibly.
}
\description{
Print method for multiple_comparisons
}
\examples{
dat.aov <- aov(Petal.Width ~ Species, data = iris)
output <- multiple_comparisons(dat.aov, classify = "Species")
print(output)
}
\seealso{
\code{\link[=multiple_comparisons]{multiple_comparisons()}}
}
