% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_methods.R
\name{process_aliased}
\alias{process_aliased}
\title{Process aliased treatments in predictions}
\usage{
process_aliased(
  pp,
  sed,
  classify,
  exclude_cols = c("predicted.value", "std.error", "df", "Names")
)
}
\arguments{
\item{pp}{Data frame of predictions}

\item{sed}{Standard error of differences matrix}

\item{classify}{Name of predictor variable}

\item{exclude_cols}{Column names to exclude when processing aliased names}
}
\value{
List containing processed predictions, sed matrix and aliased names
}
\description{
Process aliased treatments in predictions
}
\keyword{internal}
