% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resplot.R
\name{resplot}
\alias{resplot}
\alias{resplt}
\title{Residual plots of linear models.}
\usage{
resplot(
  model.obj,
  shapiro = TRUE,
  call = FALSE,
  label.size = 10,
  axes.size = 10,
  call.size = 9,
  mod.obj
)

resplt(
  model.obj,
  shapiro = TRUE,
  call = FALSE,
  label.size = 10,
  axes.size = 10,
  call.size = 9,
  mod.obj
)
}
\arguments{
\item{model.obj}{An \code{aov}, \code{lm}, \code{lme} (\code{\link[nlme:lme]{nlme::lme()}}), \code{lmerMod} (\code{\link[lme4:lmer]{lme4::lmer()}}), \code{asreml} or \code{mmer} (sommer) model object.}

\item{shapiro}{(Logical) Display the Shapiro-Wilks test of normality on the plot?}

\item{call}{(Logical) Display the model call on the plot?}

\item{label.size}{A numeric value for the size of the label (A,B,C) font point size.}

\item{axes.size}{A numeric value for the size of the axes label font size in points.}

\item{call.size}{A numeric value for the size of the model displayed on the plot.}

\item{mod.obj}{Deprecated to be consistent with other functions. Please use \code{model.obj} instead.}
}
\value{
A list containing ggplot2 objects which are diagnostic plots.
}
\description{
Produces plots of residuals for assumption checking of linear (mixed) model.
}
\examples{
dat.aov <- aov(Petal.Length ~ Petal.Width, data = iris)
resplot(dat.aov)
resplt(dat.aov, call = TRUE)
}
